+++ src/app/globals.css
@@ -102,7 +102,6 @@
 
   ul,
   li {
-    list-style: none;
     text-decoration: none;
     padding-left: 0px;
     margin: 0;
